// src/app/types/Canje/TItemReport.ts
export type TItemReport = {
    numero_documento?: string
    nombre_colaborador?: string
    fecha_otorgamiento?: string
    fecha_inicio_subsidio?: string
    fecha_final_subsidio?: string
    total_dias?: string
    fecha_maxima_canje?: string
    nombre_tipodescansomedico?: string
    nombre_tipocontingencia?: string
    mes_devengado?: string
}